moved changelog tests to separate file
